requirements.txt
alembic==1.13.1
alembic_postgresql_enum==1.1.2
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
click==8.1.7
ecdsa==0.18.0
exceptiongroup==1.2.0
fastapi==0.110.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
Mako==1.3.2
MarkupSafe==2.1.5
packaging==23.2
pluggy==1.4.0
psycopg2-binary==2.9.9
pur==7.3.1
pyasn1==0.5.1
pydantic==2.6.3
pydantic_core==2.16.3
pytest==8.0.2
python-dotenv==1.0.1
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
sqlalchemy2-stubs==0.0.2a37
sqlmodel==0.0.16
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.27.1
passlib[bcrypt]

